mute done

+++ app/api/src/module/chat/gateway/chatWs.gateway.ts
@@ -207,13 +207,15 @@ export class ChatWsGateway implements OnGatewayConnection, OnGatewayDisconnect {
             }
         }
         else if (payload.action === 'mute') {
-            // await this.chatWsService.muteUser(payload.reciever, payload.reciever)
+            await this.chatWsService.muteUser(payload.reciever, payload.user, payload.sender);
+            await this.setupSpecialMessage(payload.sender, payload.reciever, `${payload.sender} muted ${payload.user}`)
         }
         else if (payload.action === 'ban') {
             await this.chatWsService.banUser(payload.reciever, payload.reciever, payload.sender)
             const clientSocket = this.getSocket(payload.user)
             await this.setupSpecialMessage(payload.sender, payload.reciever, `${payload.sender} banned ${payload.user}`)
-            if (clientSocket) clientSocket.leave(payload.reciever)
+            if (clientSocket)
+                clientSocket.leave(payload.reciever)
         }
         else if (payload.action === 'reset') {
             // await this.chatWsService.resetUser(payload.reciever, payload.reciever)
@@ -233,7 +235,7 @@ export class ChatWsGateway implements OnGatewayConnection, OnGatewayDisconnect {
             return this.socketError('Invalid reciever')
 
         if (!(await this.chatWsService.isUserNormal(payload.reciever, payload.sender)))
-            return this.socketError('User is not in the chat room')
+            return this.socketError('User is not normal in the chat room')
         
         this.chatService.createMessage(payload.sender, payload.reciever, payload.message)
 
@@ -251,6 +253,9 @@ export class ChatWsGateway implements OnGatewayConnection, OnGatewayDisconnect {
             return this.socketError('User not found')
         if (!(await this.chatService.validateChatRoom(payload.reciever, payload.sender)))
             return this.socketError('Invalid reciever')
+
+        if (!(await this.chatWsService.isUserNormal(payload.reciever, payload.sender)))
+            return this.socketError('User is not normal in the chat room')
 
         this.chatService.deleteMessage(payload.sender, payload.reciever, payload.message_id)
